requirements.txt
attrs
cryptography
typing_extensions